Remove the PyCall/SymPy stuff
