+++ Project.toml
@@ -23,7 +23,6 @@ MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 MonteCarloMeasurements = "0987c9cc-fe09-11e8-30f0-b96dd679fdca"
 NamedTupleTools = "d9ec5142-1e00-5aa0-9d6a-321866360f50"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
-PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
@@ -36,7 +35,6 @@ SimplePosets = "b2aef97b-4721-5af9-b440-0bad754dc5ba"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
-SymPy = "24249f21-da20-56a4-8eb1-6a02cf4ae2e6"
 TransformVariables = "84d833dd-6860-57f9-a1a7-6da5db126cff"
 
 [compat]
@@ -56,7 +54,6 @@ MLStyle = "0.3, 0.4"
 MacroTools = "0.5"
 MonteCarloMeasurements = "0.6,0.7, 0.8, 0.9"
 NamedTupleTools = "0.12, 0.13"
-PyCall = "1.91"
 RecipesBase = "0.7,0.8, 1.0"
 Reexport = "0.2"
 Requires = "1.0"
@@ -67,12 +64,10 @@ SimplePartitions = "0.2, 0.3"
 SimplePosets = "0.0, 0.1"
 SpecialFunctions = "0.10"
 StatsFuns = "0.9"
-SymPy = "1.0"
 TransformVariables = "0.3"
 julia = "1.3"
 
 [extras]
-BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
@@ -81,4 +76,4 @@ StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["BenchmarkTools", "Documenter", "Literate", "Plots", "Random", "StableRNGs", "Test"]
+test = ["Documenter", "Literate", "Plots", "Random", "StableRNGs", "Test"]
