+++ src/Soss.jl
@@ -46,8 +46,9 @@ include("transforms/utils.jl")
 include("transforms/basictransforms.jl")
 include("transforms/withdistributions.jl")
 
-include("symbolic/symbolic.jl")
 include("symbolic/codegen.jl")
+# include("symbolic/symbolic.jl") # TODO: move this to the SossSymPy.jl package
+# include("symbolic/codegen-sympy.jl") # TODO: move this to the SossSymPy.jl package
 
 include("particles.jl")
 include("plots.jl")
