+++ test/linear-model.jl
@@ -32,7 +32,7 @@ end;
 jointdist = m2(X=X)
 
 @testset "logpdf, symlogpdf" begin
-    @test_nowarn symlogpdf(m2).evalf(3)
+    # @test_nowarn symlogpdf(m2).evalf(3)
     @test logpdf(jointdist, truth) ≈ -28.551921801470908
-    @test logpdf(jointdist, truth, codegen) ≈ -28.551921801470904
+    # @test logpdf(jointdist, truth, codegen) ≈ -28.551921801470904
 end
